services/web/requirements.txt
blinker==1.7.0
certifi==2023.5.7
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
DateTime==5.4
dnspython==2.5.0
email-validator==2.1.0.post1
flake8==6.1.0
Flask==3.0.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
git-filter-repo==2.38.0
greenlet==3.0.3
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
mysql-connector-python==8.3.0
numpy==1.26.3
packaging==23.2
pdfkit==1.0.0
pillow==10.2.0
protobuf==4.25.2
pycodestyle==2.11.1
pyflakes==3.1.0
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
reportlab==4.0.9
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.23
typing_extensions==4.9.0
urllib3==2.2.0
Werkzeug==3.0.1
WTForms==3.1.2
zope.interface==6.1
